improve file structure.
