+++ include/minishell.h
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   minishell.h                                        :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: kmiyazaw <kmiyazaw@student.42.fr>          +#+  +:+       +#+        */
+/*   By: miyazawa.kai.0823 <miyazawa.kai.0823@st    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/10/13 16:52:07 by kmiyazaw          #+#    #+#             */
-/*   Updated: 2023/10/13 19:16:39 by kmiyazaw         ###   ########.fr       */
+/*   Updated: 2023/10/15 15:50:19 by miyazawa.ka      ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -21,5 +21,21 @@
 
 # include <signal.h>
 # include <unistd.h>
+# include <sys/stat.h>
+#include <dirent.h>
+
+typedef struct s_data
+{
+	int			ac;
+	char		**av;
+	char		**envp;
+
+
+}				t_data;
+
+void	handler(int sig);
+void	minishell(void);
+
+
 
 #endif
